restructure proto pkgs
