+++ Makefile
@@ -94,7 +94,8 @@ buf-deps:
 .PHONY: proto-generate ## Compile protobuf to pb files
 proto-generate:
 	@echo "\n + Generating pb language bindings\n"
-	@buf generate --path ./metro-proto/metro/proto
+	@buf generate --path ./metro-proto/metro/pubsub
+	@buf generate --path ./metro-proto/metro/admin
 	# Generate static assets for OpenAPI UI
 	@statik -m -f -src third_party/OpenAPI/
