+++ internal/project/validation.go
@@ -7,7 +7,7 @@ import (
 	"strings"
 
 	"github.com/razorpay/metro/pkg/logger"
-	metrov1 "github.com/razorpay/metro/rpc/proto/v1"
+	metrov1 "github.com/razorpay/metro/rpc/pubsub/v1"
 )
 
 var idRegex *regexp.Regexp
