+++ internal/publisher/publisher.go
@@ -3,7 +3,7 @@ package publisher
 import (
 	"context"
 
-	metrov1 "github.com/razorpay/metro/rpc/proto/v1"
+	metrov1 "github.com/razorpay/metro/rpc/pubsub/v1"
 )
 
 // IPublisher is interface over high level publisher
