+++ internal/subscription/validation.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/razorpay/metro/internal/merror"
 	"github.com/razorpay/metro/internal/topic"
-	metrov1 "github.com/razorpay/metro/rpc/proto/v1"
+	metrov1 "github.com/razorpay/metro/rpc/pubsub/v1"
 )
 
 var subscriptionNameRegex *regexp.Regexp
