+++ internal/topic/validation_test.go
@@ -6,7 +6,8 @@ import (
 	"context"
 	"testing"
 
-	metrov1 "github.com/razorpay/metro/rpc/proto/v1"
+	adminv1 "github.com/razorpay/metro/rpc/admin/v1"
+	metrov1 "github.com/razorpay/metro/rpc/pubsub/v1"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -54,7 +55,7 @@ func TestValidation_GetValidatedAdminModel(t *testing.T) {
 
 	topic := "projects/test-project/topics/test-topic"
 	numPartitions := 5
-	m, err := GetValidatedAdminModel(ctx, &metrov1.AdminTopic{
+	m, err := GetValidatedAdminModel(ctx, &adminv1.AdminTopic{
 		Name:          topic,
 		Labels:        nil,
 		NumPartitions: int32(numPartitions),
@@ -69,7 +70,7 @@ func TestValidation_GetValidatedAdminModel_Error(t *testing.T) {
 
 	topic := "projects/test-project/topics/test-topic"
 	numPartitions := -1 // wrong value
-	_, err := GetValidatedAdminModel(ctx, &metrov1.AdminTopic{
+	_, err := GetValidatedAdminModel(ctx, &adminv1.AdminTopic{
 		Name:          topic,
 		Labels:        nil,
 		NumPartitions: int32(numPartitions),
