+++ unit-test.exclusions
@@ -1,4 +1,5 @@
 # Path of packages to exclude from unit testing. This has to be a sorted list.
 github.com/razorpay/metro/cmd/service
 github.com/razorpay/metro/cmd/service/metro
-github.com/razorpay/metro/rpc/proto/v1
+github.com/razorpay/metro/rpc/pubsub/v1
+github.com/razorpay/metro/rpc/admin/v1
